If knot of type IMAGE gets deleted, then its image gets delted too

+++ src/components/index/Graph.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 
 import { KeycloakInstance } from "keycloak-js";
-import Api, { Intent, Knot } from "metamind-client";
+import Api, { Intent, Knot, KnotType } from "metamind-client";
 import { connect } from "react-redux";
 import { Dispatch } from "redux";
 import { IStoreState } from "src/types";
@@ -334,7 +334,10 @@ class Graph extends React.Component<IProps, IState> {
    if ( this.props.knots ) {
      const knot = this.props.knots.find((k) => k.id === viewNode.id);
      const knotContent = knot ? knot.content : "";
-     await fetch(`${process.env.REACT_APP_API_BASE_PATH}/images/${knotContent}`, {method: "delete"});
+     const knotType = knot ? knot.type : "";
+     if ( knotType === KnotType.IMAGE ) {
+       await fetch(`${process.env.REACT_APP_API_BASE_PATH}/images/${knotContent}`, {method: "delete"});
+     }
    }
    await Api.getKnotsService(this.props.keycloak ? this.props.keycloak.token! : "").deleteKnot(this.props.storyId, viewNode.id);
